// src/types/table.ts
export interface Billboard {
  id: string;
  title: string;
  artist: string;
  type: string;
  peak: string;
  wks: string;
}

export interface IMDb {
  rank: string;
  title: string;
  year: string;
  length: string;
  age: string;
  rating: string;
  num_votes: string;
}

export interface IMDb {
  rank: string;
  title: string;
  year: string;
  length: string;
  age: string;
  rating: string;
  num_votes: string;
}

export interface YouTube {
  title: string;
  published_at: string;
  view_count: string;
  thumbnail_url: string;
  video_url: string;
  isoDateString: string;
}

export interface Netflix {
  rows: string;
  title: string;
  episode: string;
  wks: string;
  views: number;
}

export interface Netflix {
  rows: string;
  title: string;
  episode: string;
  wks: string;
  views: number;
}

export interface NYT {
  title: string;
  author: string;
  wks: string;
  desc: string;
}